move sunset logic to get-products and make covers fetching dynamic
