+++ scripts/v2-migration/products/get-products.js
@@ -5,6 +5,70 @@ const fetch = require('node-fetch');
 
 const outputDir = path.join(__dirname, 'output');
 
+const gracePeriod = {
+  protocol: 30,
+  custodian: 120,
+  token: 14,
+};
+
+function getExpiryTime(products, cover) {
+  const product = products[cover.address];
+  if (!product) {
+    console.log(`Product not found for ${cover.address}`);
+    return -1;
+  }
+
+  // Cover expiration time + grace period
+  const endDate = new Date(cover.end_time);
+  return endDate.getTime() + gracePeriod[product.type] * 24 * 60 * 60 * 1000;
+}
+
+function decode(buf) {
+  const decoder = new TextDecoder('utf-8');
+  return decoder.decode(buf);
+}
+
+const getSunsetProducts = async () => {
+  const url = 'https://nexustracker.io/all_covers';
+  const covers = JSON.parse(
+    decode(await fetch(url, { headers: { 'Content-Type': 'application/json' } }).then(x => x.arrayBuffer())),
+  );
+
+  const products = await fetch('https://api.nexusmutual.io/coverables/contracts.json').then(r => r.json());
+
+  const productsWithLowerCasedKeys = {};
+  for (const key in products) {
+    productsWithLowerCasedKeys[key.toLowerCase()] = products[key];
+  }
+
+  const latestCoverExpiryPerProduct = {};
+  for (const cover of covers) {
+    const expiryTime = getExpiryTime(productsWithLowerCasedKeys, cover);
+
+    if (expiryTime === -1) {
+      // product not found
+      continue;
+    }
+
+    const latestExpiryTime = latestCoverExpiryPerProduct[cover.address];
+    if (latestExpiryTime) {
+      latestCoverExpiryPerProduct[cover.address] = expiryTime > latestExpiryTime ? expiryTime : latestExpiryTime;
+    } else {
+      latestCoverExpiryPerProduct[cover.address] = expiryTime;
+    }
+  }
+
+  const now = new Date().getTime();
+  const sunsetProductKeys = Object.keys(products).filter(p => {
+    if (!latestCoverExpiryPerProduct[p.toLowerCase()]) {
+      return true;
+    }
+    return latestCoverExpiryPerProduct[p.toLowerCase()] < now;
+  });
+
+  return sunsetProductKeys;
+};
+
 const getProductsContract = products => `// SPDX-License-Identifier: GPL-3.0-only
 
 pragma solidity >=0.5.0;
@@ -39,7 +103,7 @@ const main = async () => {
   // Add deprecated products that are not sunset
   const deprecatedProductAddresses = v1ProductAddresses.filter(k => v1Products[k].deprecated);
   // TODO move the logic here rather than reading from an output file
-  const sunsetProducts = require(path.join(__dirname, 'output/sunsetProducts.json'));
+  const sunsetProducts = await getSunsetProducts();
   v2ProductAddresses.push(...deprecatedProductAddresses.filter(k => sunsetProducts.indexOf(k) === -1));
 
   console.log(`Total V2 products: ${v2ProductAddresses.length}`);
